# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger in Laravel with Passport Auth Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "bosavras@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Return users",
                "description": "Get users",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::index",
                "responses": {
                    "200": {
                        "description": "Array of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Id",
                                                "type": "integer",
                                                "nullable": false
                                            },
                                            "First name": {
                                                "description": "First Name",
                                                "type": "string",
                                                "nullable": false
                                            },
                                            "Last name": {
                                                "description": "Last Name",
                                                "type": "string",
                                                "nullable": false
                                            },
                                            "email": {
                                                "description": "User email",
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "company": {
                                                "description": "User company",
                                                "type": "object",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user by id",
                "description": "Update user by id",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Update user body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "First name": {
                                        "description": "First Name",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "Last name": {
                                        "description": "Last Name",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Id",
                                            "type": "integer",
                                            "nullable": false
                                        },
                                        "First name": {
                                            "description": "First Name",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "Last name": {
                                            "description": "Last Name",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "email": {
                                            "description": "User email",
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation exceptions"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "description": "Create user",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::create",
                "requestBody": {
                    "description": "New body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "First name": {
                                        "description": "First Name",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "Last name": {
                                        "description": "Last Name",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validation exceptions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Id",
                                            "type": "integer",
                                            "nullable": false
                                        },
                                        "First name": {
                                            "description": "First Name",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "Last name": {
                                            "description": "Last Name",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "email": {
                                            "description": "User email",
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Return user by id",
                "description": "Get user by id",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "event",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Id",
                                            "type": "integer",
                                            "nullable": false
                                        },
                                        "First name": {
                                            "description": "First Name",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "Last name": {
                                            "description": "Last Name",
                                            "type": "string",
                                            "nullable": false
                                        },
                                        "email": {
                                            "description": "User email",
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user by id",
                "description": "Delete user by id",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/api/v1/user/company/{id}": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Appoint current user to id",
                "description": "Appoint current user to id",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\UserController::appointCompanyById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "company id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointed"
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    ]
}